Create all Classes-Files

+++ CG2-Warmup-Java/src/cg2/warmup/Checkerboard.java
@@ -6,8 +6,8 @@ import cg2.vecmath.Color;
 public class Checkerboard implements Painter {
 	
 	public Color pixelColorAt(final int i, final int j, final int width, final int height) {
-		final double rectwidth =  Math.floor((double) i / (double) (width / 8.0));
-		final double rectheight =  Math.floor((double) j / (double) (height / 8.0));
+		final double rectwidth =  Math.floor( i /  (width / 8.0));
+		final double rectheight =  Math.floor( j /  (height / 8.0));
 		
 		if(rectwidth % 2 + rectheight % 2 == 1) {
 			return new Color(0, 0, 0);
